Remove cache.py
